Rename scrollbar util
